Add files via upload
This code will help you to understand the numpy and pandas library basics in python. pandas is a fast, powerful, flexible and easy to use open source data analysis and manipulation tool,built on top of the Python programming language. Numpy  is basically used for working with arrays in python.